chaged long_int to size_t

+++ src/rc/access.c
@@ -54,11 +54,11 @@
 static HtmlTagDef file_tag_def = NULL;
 
 static int attach_archive(RcArchive rca);
-static int attach_archive_mem(RcArchive rca, char *mem, long int mem_size);
+static int attach_archive_mem(RcArchive rca, char *mem, size_t mem_size);
 
 /* MOD-OL olsky OL-2016: patch for mem */
 RcArchive
-rc_open_archive_mem(const char *file, char *mem, long int mem_size, int flags)
+rc_open_archive_mem(const char *file, char *mem, size_t mem_size, int flags)
 { RcArchive rca = malloc(sizeof(rc_archive));
 
   if ( rca )
@@ -454,7 +454,7 @@ scan_archive(RcArchive rca)
 
 /* olsky OL-2016: patch for mem */
 static int
-attach_archive_mem(RcArchive rca, char *mem, long int mem_size)
+attach_archive_mem(RcArchive rca, char *mem, size_t mem_size)
 {
     rca->map_size  = mem_size;
     rca->size      = rca->map_size;
